graphclean: rm files and creates edges between options
The graph obtained with make graphconfig is big since previous commit on
reducing the granularity. It is bigger because of additional nodes for
files. This program creates an undirected graph with edges between options
pointing to the same file, e.g that are interfering because they are in the same
file.