scripts/kconfig/confdata: improve cmd for graphconfig in c_read
the cmd for retrieving files that contain an option has been improved. Before it
was a really tricky one with pipes and all that... Now, it uses grep's
options. It does the search only on .c, .h and Makefiles and look first only the
first match. Generating the graph is faster than the previous implementation
now.