kconfig: cfconfig: main that mutates .conf with random syms
cfconfig will pick n random bool or tristate symbols and generate one
configuration per addition of the symbols with the following form at
srctree:
    ___config<OPTION_NAME>_<given value>-<solution numer>)

All possible values of the symbols will be tested. If sym is bool and its
value is 'n', the program will find a solution for 'y'. If tristate and n,
it will generate for 'm' and 'y', and so on...

At the end, the program will print the number of draw (symbols picked, which is
the given 'n'), number of file generated and number of solutions not found by
Rangefix.

Run:
    export SRCARCH=x86 LD=$(which ld) CC=$(which gcc) srctree=$(pwd) ;\
           scripts/kconfig/cfconfig Kconfig <n>